Rename the 'About the vacancy' page
This was called 'About' in our codebase. I've renamed it to 'Vacancy', which is a more descriptive name for the page.
